pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splunk-sigma"
version = "0.1.0"
description = "Splunk app that runs Sigma detection rules natively via a custom | sigma search command"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [{ name = "Jacob Hess" }]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pyyaml>=6.0",
    "ruff>=0.1.0",
    "splunk-sdk>=2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["app/bin"]

[tool.ruff]
line-length = 100
target-version = "py37"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
